Changes http links to https
This is similar as a change set to getpelican/pelican-themes@cf3d1ed3bd47ec539993135bf193b0bead8abfe2
by @jeremyn. (Here we are also changing the link to the Github repo.)

Note that we aren't changing http IRIs identifying resources for machines,
such as for XML namespaces and for Dublin Core terms.